Projectile JBox2D shape definition _almost_ properly defined. Either the JBox2D definition or the JavaFX shape is one radius length too long.
Utility changed back to floats, because for some reason a bunch of
things started to complain.

+++ src/Game/GameWorld.java
@@ -32,8 +32,8 @@ public class GameWorld {
 		backGround = new ImageView(back);
 		pWidth = back.getWidth();
 		pHeight = back.getHeight();
-		width = (int) Utility.toWidth(pWidth);
-		height = (int) Utility.toHeight(pHeight);
+		width = (int) Utility.toWidth((float)pWidth);
+		height = (int) Utility.toHeight((float)pHeight);
 		addElements();
 	}
